LayoutTests/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-stdDeviation-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dropShadowElement.stdDeviationX.baseVal is 0
PASS dropShadowElement.stdDeviationY.baseVal is 0
PASS dropShadowElement.setStdDeviation(10, 10) did not throw exception.
PASS dropShadowElement.setStdDeviation(NaN, NaN) threw exception TypeError: The provided value is non-finite.
PASS dropShadowElement.setStdDeviation(Infinity, Infinity) threw exception TypeError: The provided value is non-finite.
PASS dropShadowElement.setStdDeviation() threw exception TypeError: Not enough arguments.
PASS dropShadowElement.setStdDeviation(0) threw exception TypeError: Not enough arguments.
PASS dropShadowElement.stdDeviationX.baseVal is 10
PASS dropShadowElement.stdDeviationY.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
